Move test_wlcs_display_server.* to the wlcs integration module
